+++ src/model/ofxGgmlModel.h
@@ -10,6 +10,7 @@ struct ofxGgmlModelInfo {
 	std::string architecture;
 	uint64_t tensorCount = 0;
 	uint64_t metadataCount = 0;
+	uint64_t layerCount = 0;
 };
 
 class ofxGgmlModel {